requirements.txt
keyboard==0.13.5
numpy==1.23.4
pandas==1.5.0
plotly==5.13.0
pykalman==0.9.5
pyserial==3.5
tqdm==4.64.1
joblib==1.2.0
matplotlib==3.6.2
scikit_learn==1.1.2
scipy==1.11.1
